refactor(browser): extract async helpers and project-format tests from app/mod
